requirements.txt
numpy==1.16
pandas
sklearn
scikit-tensor-py3
scipy==1.3
networkx
argparse
termcolor
pyyaml
pytest